01._Assignments/03._Server_To_Server/02._Python/pyproject.toml
[project]
name = "02-python"
version = "0.1.0"
description = ""
authors = [
    {name = "realkoder",email = "alexanderbtcc@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "uvicorn (>=0.34.0,<0.35.0)",
    "fastapi (>=0.115.8,<0.116.0)",
    "pyyaml (>=6.0.2,<7.0.0)"
]

[tool.poetry]
packages = [{include = "02"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
